% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugsSideEffect.R
\name{drugsSideEffect}
\alias{drugsSideEffect}
\title{Ministry of Food and Drug Safety, Drugs side effect information.}
\usage{
drugsSideEffect(key)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}
}
\description{
drugsSideEffect function import the side effects of drugs.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example 1 searching by localeCode.
 data <- drugsSideEffect(key)

}
