% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episJoint.R
\name{episJoint}
\alias{episJoint}
\title{Korea Agency of Education, Promotion and Informaion Service in Food, Agriculture, Forestry and Fisheries. Price of food, agriculture, forestry and fisheries from joint market.}
\usage{
episJoint(key, date, verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{date}{character value. date which expressed like YYYYMMDD. no default.}

\item{verbose}{logical value. If TRUE, show process bar. Default is set as FALSE.}
}
\description{
episJoint function import food, agriculture, forestry and fisheries price in joint market.
  This function need API key issued from 'data.mafra.go.kr', not 'data.go.kr'.
}
\details{
The date arguments should be set as YYYYMMDD type like '20181203'.
  The API key issued from 'data.mafra.go.kr' is needed, not 'data.go.kr'.
}
\examples{
 key <- 'your key issued from data.mafra.go.kr'
 data <- episJoint(key, date = '20191203')

}
