% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episPrice.R
\name{episPrice}
\alias{episPrice}
\title{Korea Agency of Education, Promotion and Informaion Service in Food, Agriculture, Forestry and Fisheries. Price of food, agriculture, forestry and fisheries.}
\usage{
episPrice(key, date, items, verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{date}{character value. date which expressed like YYYYMMDD. no default.}

\item{items}{numeric value. item code in epis_items data.}

\item{verbose}{logical value. If TRUE, show process bar. Default is set as FALSE.}
}
\description{
episPrice function import price of food, agriculture, forestry and fisheries.
  It provide similar data to atPrice function but have different survey source.
}
\details{
The date arguments should be set as YYYYMMDD type like '20181203'.
  The survey of food, agriculture, forestry and fisheries is not excuted on weekend.
  So if you insert sunday, saturday or holiday date in date argument, there will be no return.
}
\examples{
 key <- 'your key issued from data.go.kr'
 epis_items <- datagokR::epis_items

 data <- episPrice(key, date = '20191203', items = head(unique(epis_items$itemCode)))

}
