% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_xml.R
\name{find_xml}
\alias{find_xml}
\title{Find text from xml2::read_xml object and if not exist, return NA.}
\usage{
find_xml(xml, tag)
}
\arguments{
\item{xml}{result of xml2::read_xml function.}

\item{tag}{tag which you want to find.}
}
\description{
Find text from xml2::read_xml object and if not exist, return NA.
}
