% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_xmlList.R
\name{find_xmlList}
\alias{find_xmlList}
\title{Find text from XML::xmlToList object and if not exist, return NA.}
\usage{
find_xmlList(xml, tag, type = "char")
}
\arguments{
\item{xml}{result of xml2::read_xml function.}

\item{tag}{tag which you want to find.}

\item{type}{decide class of result between character(char) and numeric(num).}
}
\description{
Find text from XML::xmlToList object and if not exist, return NA.
}
