% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiraCancerAssess.R
\name{hiraCancerAssess}
\alias{hiraCancerAssess}
\title{Health Insurance Review & Assessment Service, Cancer Surgery Mortality Assessment.}
\usage{
hiraCancerAssess(key)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}
}
\description{
hiraCancerAssess function import hosipital assessment grade data about liver, gastric cancer surgery mortality.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- hiraCancerAssess(key)

}
