% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kma_branches.R
\docType{data}
\name{kma_branches}
\alias{kma_branches}
\title{Data for kmaASOS function.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4626 rows and 7 columns.}
\usage{
kma_branches
}
\description{
A dataframe containing information about branches with observe weather(ASOS) data.
 It is tbl_df type object, so "dplyr" library is recommended.
 The variables are as follow.
}
\details{
\itemize{
  \item code. numeric branch code.
  \item startDate. date of the branch is opened.
  \item endDate. date of the branch is expired. If empty, it means that the brach is still operated.
  \item name. location.
  \item long. longitude code.
  \item lat. latitude code.
 }

@docType data

@usage data(kma_branches)

@format A dplyr Dataframe with 7 columns and 4,626 rows

@keywords datasets

@name kma_branches

@examples
 data(kma_branches)
}
\keyword{datasets}
