% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kma_lifeIndex_type_check.R
\docType{data}
\name{kma_lifeIndex_type_check}
\alias{kma_lifeIndex_type_check}
\title{Data for kmaLifeIndex function.}
\format{An object of class \code{matrix} with 8 rows and 12 columns.}
\usage{
kma_lifeIndex_type_check
}
\description{
A matrix containing TRUE or FALSE which means the availability of indeces for each month.
}
\details{
@docType data

@usage data(kma_lifeIndex_type_check)

@format A matrix of 8 by 12

@keywords datasets

@name kma_lifeIndex_type_check

@examples
 data(kma_lifeIndex_type_check)
}
\keyword{datasets}
