% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kma_lifeIndex_urlType.R
\docType{data}
\name{kma_lifeIndex_urlType}
\alias{kma_lifeIndex_urlType}
\title{Data for kmaLifeIndex function.}
\format{An object of class \code{character} of length 8.}
\usage{
kma_lifeIndex_urlType
}
\description{
A character vector for matching "type" argument in kmaLifeIndex function and API url.
}
\details{
@docType data

@usage data(kma_lifeIndex_urlType)

@format A named vector.

@keywords datasets

@name kma_lifeIndex_urlType

@examples
 data(kma_lifeIndex_urlType)
}
\keyword{datasets}
