% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kma_locale_code.R
\docType{data}
\name{kma_locale_code}
\alias{kma_locale_code}
\title{Data for kmaHealthIndex and kmaLifeIndex function.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3772 rows and 4 columns.}
\usage{
kma_locale_code
}
\description{
A dataset contain the legal location code and the name of locations.
 It is tbl_df type object so "dplyr" library is recommended.
 The variables are as follow.
}
\details{
\itemize{
  \item code. five-digit legal location code.
  \item name1. korean location name for Si/Do.
  \item name2. korean location name for Si/Gu.
  \item name3. korean location name for Dong.
 }

@docType data

@usage data(kma_locale_code)

@format A tbl_df type data frame with 3,772 rows and 4 columns

@keywords datasets

@name kma_locale_code

@source https://www.code.go.kr/index.do

@examples
 data(kma_locale_code)
}
\keyword{datasets}
