% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmaASOS.R
\name{kmaASOS}
\alias{kmaASOS}
\title{Korea Meteorological Administration, Automated Synoptic Observing System(ASOS) data of day.}
\usage{
kmaASOS(key, branchCode = NULL, fromDate = NULL, toDate = NULL,
  slow = F, errorCheck = T, verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <https://data.kma.go.kr>. Note that it's not from <data.go.kr>. no default.}

\item{branchCode}{numeric value. the codes of branches. If insert "all", import all branches data. See kma_branch data for details.}

\item{fromDate}{date value. 8-digits date which means starting date.}

\item{toDate}{date value. 8-digits date which means end date.}

\item{slow}{logical value. if TRUE, give sleep inbetween importing. default is TRUE.}

\item{verbose}{logical value. if TRUE, provide process bar. Default value set as false.}

\item{viz}{logical value. if TRUE, provide simple 2d visualization result. x: date, y: mean index.}
}
\value{
data.frame and visualization.
}
\description{
kmaASOS function import Korea Moteorological Administration's ASOS data of day.
}
\details{
ASOS is an observation conducted at the same time at all
 branches at a fixed time to understand the synoptic weather.
 Normaly, it refers to the size of anticyclone and cyclone or daily weather on weather map.\cr
 See more details about each data from <https://data.kma.go.kr/data/grnd/selectAsosRltmList.do> \cr
 The date argument range is from 1904-01-01 to previous day.\cr
 It should be noticed that this function needs the API key issued from https://data.kma.go.kr, not from https://data.go.kr.\cr
 Importing data from KMA is very unstable that it is recommended to checking error urls in result.
}
\examples{
 # example 1 searching by branchCode.
 data <- kmaASOS(key, branchCode = c(108), fromDate = as.date("2010-01-01"),
                 toDate = as.date("2010-01-10"), slow = T)

}
