% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmaHealthIndex.R
\name{kmaHealthIndex}
\alias{kmaHealthIndex}
\title{Korea Meteorological Administration, The health index.}
\usage{
kmaHealthIndex(key, time = c(6, 18), localeCode = NULL,
  localeName = NULL, type, slow = T, viz = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{time}{numeric value. The time when the data is generated.}

\item{localeCode}{numeric value. SiGunGu code which means legal area. one of localeCode or localeName should be inserted.}

\item{localeName}{character value. SiGunGu name wich means legal area. one of localeCode or localeName should be inserted. It should be Korean.}

\item{type}{character value. decide the type of index. it should be one of
"Asthma", "Brain", "Skin", "FlowerWoody", "FlowerPine", "FlowerWeeds", "Infl" or "possible". see details.}

\item{slow}{logical value. if TRUE, give sleep inbetween importing. default is TRUE.}

\item{viz}{logical value. if TRUE, provide simple 2d visualization result. x: date, y: mean index.}
}
\value{
data.frame and visualization.
}
\description{
kmaLifeIndex function import the health index data last 2 days. The function also provide simple visualization using plotly.
}
\details{
kmaLifeIndex function import eight index value which calculated by Korea Meteorological Administration and related to public's life.
 Explanation about "type" as follow.\cr
 "Asthma" = Possibility index of asthma or lung disease. It is povided the whole year.\cr
 "Brain" = Possibility index of stroke. It is povided the whole year.\cr
 "Skin" = Possibility index of skin disease. It is povided the whole year.\cr
 "FlowerWoody" = Risk index of oak tree's pollen. It is povided from Apr to May.\cr
 "FlowerPine" = Risk index of pine tree's pollen. It is povided from Apr to May.\cr
 "FlowerWeeds" = Risk index of weed's pollen. It is povided from Sep to Oct.\cr
 "Infl" = Possibility index of influenza. It is povided from Sep to Apr.\cr
 "possible" = All index which affordable now.\cr
 See more details about each index from <http://www.weather.go.kr/weather/lifenindustry/life_jisu.jsp#>
}
\examples{
 # example 1 searching by localeCode.
 data <- kmaHealthIndex(key, time = seq(0, 21, 3), localeCode = c(4111100000, 4111156600),
                        type = "Asthma", slow = T)

 # example 2 searching by localeName
 data <- kmaHealthIndex(key, time = seq(0, 21, 3),
                        localeName = c("수원"), type = "possible", slow = T)

}
