% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmaLifeIndex.R
\name{kmaLifeIndex}
\alias{kmaLifeIndex}
\title{Korea Meteorological Administration, The life weather index.}
\usage{
kmaLifeIndex(key, time = seq(0, 21, 3), localeCode = NULL,
  localeName = NULL, type, slow = T, viz = T)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{localeCode}{numeric value. SiGunGu code which means legal area. one of localeCode or localeName should be inserted.}

\item{localeName}{character value. SiGunGu name wich means legal area. one of localeCode or localeName should be inserted.}

\item{type}{character value. decide the type of index. it should be one of "fp", "st", "hi", "di", "ui", "fb", "ap", "sh" or "possible". see details.}

\item{slow}{logical value. if TRUE, give sleep inbetween importing. default is TRUE.}

\item{viz}{logical value. if TRUE, provide simple 2d visualization result. x: date, y: mean index.}
}
\value{
data.frame and visualization.
}
\description{
kmaLifeIndex function import the life weather index data last 2 days. The function also provide simple visualization using plotly.
}
\details{
kmaLifeIndex function import eight index value which calculated by Korea Meteorological Administration and related to public's life.
 Explanation about "type" as follow.
 "fp" = Food poison index. It is povided from Jan to Dec.
 "st" = Sensory temperature index. It is povided from Nov to Mar.
 "hi" = Heat index. It is povided from Jun to Sep.
 "di" = Discomport index. It is povided from Jun to Sep.
 "ui" = Ultraviolet index. It is povided from Mar to Nov.
 "fb" = Freezing burst index. It is povided from Dec to Feb.
 "ap" = Atmospheric dispersion index. It is povided from Nov to May.
 "sh" = Sensory heat index. It is povided from May to Sep.
 "possible" = All index which affordable now.
 See more details about each index from <http://www.weather.go.kr/weather/lifenindustry/life_jisu.jsp#>
}
\examples{
 # example 1 searching by localeCode.
 data <- molitRealTrade(key = "my_key", year = 2018, month = 1, localeCode = 11110,
                        houseType = "apart", tradeType = "trade", slow = T, viz = F)

 # example 2 searching by localeName
 data <- molitRealTrade(key = "my_key", year = 2018, month = 1:6, localeName = "서울",
                        houseType = "apart", tradeType = "rent", slow = F, viz = T)
}
