% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livestockDisease.R
\name{livestockDisease}
\alias{livestockDisease}
\title{Ministry of Agriculture Food and Rural Affairs, Livestock's disease information.}
\usage{
livestockDisease(key, fromDate = NULL, toDate = NULL, verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <data.mafra.go.kr>. no default.}

\item{fromDate}{date value. 8-digits date that start of searching period.}

\item{toDate}{date value. 8-digits date that end of searching period.}

\item{verbose}{logical value. if TRUE, provide process bar. Default value set as false.}
}
\description{
livestockDisease function import data about livestock's disease information.
  It provide name of disease, location of farm, the number of livestocks which get disease, and so on.
  This function need API key issued from 'data.mafra.go.kr', not 'data.go.kr'.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- livestockDisease(key, as.Date('2019-10-01'), as.Date('2019-10-20'), verbose = T)

}
