% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meDust.R
\name{meDust}
\alias{meDust}
\title{Ministry of Environment, Real-time Particulate Matter data.}
\usage{
meDust(key, localName = enc2utf8("서울"), condition = "HOUR",
  verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{localName}{character value. SiDo name in Korean character. Default is set as enc2utf8('서울')}

\item{condition}{character value. Refer details. Default is set as 'HOUR'}

\item{verbose}{logical value. If TRUE, show process bar. Default is set as FALSE.}
}
\description{
meDust function import real-time air polution data like so2, co, o3, no2, pm10, pm25.
}
\details{
The Contidion arguments should be set as "HOUR" or "DAILY".
  If it is set 'HOUR', function import near window data.
  If it is set 'DAILY', function import data from now to -24h.
}
\examples{
 key <- 'your key issued from data.go.kr'
 data <- meDust(key)

}
