% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molitCommerce.R
\name{molitCommerce}
\alias{molitCommerce}
\title{Ministry of Land Infrastructure and Transport, real estate for Commerce transaction data.}
\usage{
molitCommerce(key, year, month = NULL, localeCode = NULL,
  localeName = NULL, slow = F, viz = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>}

\item{year}{numeric value. the year of real trade}

\item{month}{numeric value. the month of real trade. default is NULL}

\item{localeCode}{numeric value. SiGunGu code which means legal area.}

\item{localeName}{character value. SiGunGu name wich means legal area. It should be Korean.}

\item{slow}{logical value. if TRUE, give sleep inbetween importing. default is FALSE}

\item{viz}{logical value. if TRUE, provide simple 3d visualization result. x: date, y: price, z: the number of contract.}
}
\value{
two data.frame for meta-data and imported data.
two vectors for error urls and all urls. visualization.
}
\description{
molitCommerce function import the actual transition data of house. The function also provide simple visualization using plotly.
}
\details{
If month value is NULL, all data of the year will imported.\cr
   Between localeCode and localeName, one of these parameters should be inserted. \cr
   The localeCode parameter recommended five numeric value.
}
\examples{
 # example 1 searching by localeCode.
 data <- molitCommerce(key = "my_key", year = 2018, month = 1, localeCode = 11110, slow = T, viz = F)

 # example 2 searching by localeName
 data <- molitCommerce(key = "my_key", year = 2018, month = 1:6, localeName = enc2utf8("서울"), slow = F, viz = T)

}
