% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molitDwelling.R
\name{molitDwelling}
\alias{molitDwelling}
\title{Ministry of Land Infrastructure and Transport, real estate transaction data.}
\usage{
molitDwelling(key, year, month = NULL, localeCode = NULL,
  localeName = NULL, houseType, tradeType, slow = F, viz = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>}

\item{year}{numeric value. the year of real trade}

\item{month}{numeric value. the month of real trade. default is NULL}

\item{localeCode}{numeric value. SiGunGu code which means legal area.}

\item{localeName}{character value. SiGunGu name wich means legal area. It should be Korean.}

\item{houseType}{character value. decide the type of house. it should be one of "apart", "multi", or "detached".}

\item{tradeType}{character value. decide the type of trade. it should be one of "trade" or "rent".}

\item{slow}{logical value. if TRUE, give sleep inbetween importing. default is FALSE}

\item{viz}{logical value. if TRUE, provide simple 3d visualization result. x: date, y: price, z: the number of contract.}
}
\value{
two data.frame for meta-data and imported data.
two vectors for error urls and all urls. visualization.
}
\description{
molitDwelling function import the actual transition data of house. The function also provide simple visualization using plotly.
}
\details{
If month value is NULL, all data of the year will imported.\cr
   Between localeCode and localeName, one of these parameters should be inserted. \cr
   The localeCode parameter recommended five numeric value.
   houseType parameter means the type of house. \cr
   "apart" means Apartment, "multi" means Multiplex house, and "detached" means detached house.
}
\examples{
 # example 1 searching by localeCode.
 data <- molitDwelling(key = "my_key", year = 2018, month = 1, localeCode = 11110,
                        houseType = "apart", tradeType = "trade", slow = T, viz = F)

 # example 2 searching by localeName
 data <- molitDwelling(key = "my_key", year = 2018, month = 1:6, localeName = "서울",
                        houseType = "apart", tradeType = "rent", slow = F, viz = T)

}
