% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molitRealTrade.R
\name{molitRealTrade}
\alias{molitRealTrade}
\title{Ministry of Land Infrastructure and Transport actual transition the information}
\usage{
molitRealTrade(key, year, month = NULL, locale, houseType, tradeType,
  slow = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>}

\item{year}{numeric value. the year of real trade}

\item{month}{numeric value. the month of real trade. default is NULL}

\item{locale}{numeric value. SiGunGu code which means legal area.}

\item{houseType}{character value. decide the type of house. it should be one of "apart", "multi", or "detached".}

\item{tradeType}{character value. decide the type of trade. it should be one of "trade" or "rent".}

\item{slow}{logical value. if TURE, give sleep inbetween importing. default is FALSE}
}
\value{
data.frame
}
\description{
molitRealTrade function import the actual transition data of house on the law.
}
\details{
If month value is NULL, all data of the year will imported. The locale parameter recommended five numeric value.
}
\examples{
 data <- molitRealTrade(key = "my_key", year = 2018, month = 1, locale = 11110,
                        houseType = "apart", tradeType = "realtrade", slow = T)
}
