% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasBillProposer.R
\name{nasBillProposer}
\alias{nasBillProposer}
\title{National Assembly Secretariat, Recent proposed bill list searching by congressman name.}
\usage{
nasBillProposer(key, id = NULL, verbose = FALSE)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{verbose}{logical value. If TRUE, show process bar. Default is set as FALSE.}

\item{name}{character value. the name of congressman.}
}
\description{
nasBillProposer function import proposed bill list searching by name of congressman.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasBillProposer(key, id = 'PRC_L1E6Z0K1F2V8U1N7M3V5N1Y6V6G5U3')

}
