% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasBillSearch.R
\name{nasBillSearch}
\alias{nasBillSearch}
\title{National Assembly Secretariat, Recent proposed bill list searching by congressman name.}
\usage{
nasBillSearch(key, name = NULL)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{name}{character value. the name of congressman.}
}
\description{
nasBillSearch function import proposed bill list searching by name of congressman.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasBillSearch(key, name = enc2utf8('경대수'))

}
