% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasCongressman1.R
\name{nasCongressman1}
\alias{nasCongressman1}
\title{National Assembly Secretariat, Congressman information.}
\usage{
nasCongressman1(key)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}
}
\description{
nasCongressman1 function import brief data about congressman's name, constituency, and so on.
If you need more specific information about congressman, please use nasCongressman2 function.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasCongressman1(key)

}
