% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasCongressman2.R
\name{nasCongressman2}
\alias{nasCongressman2}
\title{National Assembly Secretariat, Congressman information.}
\usage{
nasCongressman2(key, code_dept, code_numb)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{code_dept}{character value. It is "code_dept" column which can be acheived from nasCongressman1.}

\item{code_numb}{character value. It is "code_numb" column which can be acheived from nasCongressman1.}
}
\description{
nasCongressman2 function import detailed data about congressman.
You need to know department code and unique code that can be acheived from nasCongressman1.
department code is "code_dept" and unique code is "code_numb" from nasCongressman1 function.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasCongressman2(key, code_dept = '9770931', code_numb = '2952')

}
