% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasElection.R
\name{nasElection}
\alias{nasElection}
\title{National Assembly Secretariat, election information.}
\usage{
nasElection(key)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}
}
\description{
nasElection function import detail information of election like date, id, type and so on.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasElection(key)

}
