% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasPassedBill.R
\name{nasPassedBill}
\alias{nasPassedBill}
\title{National Assembly Secretariat, Recent passed bill list.}
\usage{
nasPassedBill(key)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}
}
\description{
nasPassedBill function import passed bill list of recent(about 6 months).
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasPassedBill(key)

}
