% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasPetitionSearch.R
\name{nasPetitionSearch}
\alias{nasPetitionSearch}
\title{National Assembly Secretariat, Recent proposed petition list searching by congressman name who approving the petition.}
\usage{
nasPetitionSearch(key, name = NULL)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{name}{character value. the name of congressman who approving it.}
}
\description{
nasPetitionSearch function import proposed petition list searching by name of congressman who approving the petition.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasBillSearch(key, name = enc2utf8('경대수'))

}
