% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasPolling.R
\name{nasPolling}
\alias{nasPolling}
\title{National Assembly Secretariat, polling place data.}
\usage{
nasPolling(key, election_id = NULL, localeName = enc2utf8("서울"),
  is_pre = FALSE)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{election_id}{character value. the election id which means election date.}

\item{localeName}{character value. Korean Si/Do name.}

\item{is_pre}{boolean value. If you want to get data of early voting place, set TRUE.}
}
\description{
nasPolling function import information such as address, name and so on about polling place.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasPolling(key, election_id = '20190403',
                    localeName = enc2utf8('서울'))

}
