% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhisDiseaseForcast.R
\name{nhisDiseaseForcast}
\alias{nhisDiseaseForcast}
\title{National Health Insurance Service, The Disease forcasting data.}
\usage{
nhisDiseaseForcast(key, localeCode = NULL, localeName = NULL, type,
  slow = F, viz = F, verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{localeCode}{numeric value. SiGunGu code which means legal area. one of localeCode or localeName should be inserted.}

\item{localeName}{character value. SiGunGu name wich means legal area. one of localeCode or localeName should be inserted. It should be Korean.}

\item{type}{character value. decide the type of disease. it should be one of
"Influenza", "Eye", "Food", "Asthma", "Skin", or "All". see details.}

\item{slow}{logical value. if TRUE, give sleep inbetween importing. default is TRUE.}

\item{viz}{logical value. if TRUE, provide simple 2d visualization result. x: date, y: mean index.}

\item{verbose}{logical value. if TRUE, provide process bar. Default value set as false.}
}
\value{
data.frame and visualization.
}
\description{
nhisDiseaseForcast function import the disease forcasting data about 3 days.
}
\details{
nhisDiseaseForcast function import forcasted count of treatment and risk of five diseases.\cr
 It is calculated by National Health Insurance Service.\cr
 Explanation about "type" as follow.\cr
 "Influenza" = Predicted number of patients and risk of influenza from the day to the day after tomorrow.\cr
 "Eye" = Predicted number of patients and risk of eye disease at the day.\cr
 "Food" = Predicted number of patients and risk of food poisoning from the day to the day after tomorrow.\cr
 "Asthma" = Predicted number of patients and risk of asthma at the day.\cr
 "Skin" = Predicted number of patients and risk of skin disease from the day to the day after tomorrow.\cr\cr
 localeCode argument should be first 2-digit value of SiGunGu code like "11" for Seoul or "12" for Busan.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example 1 searching by localeCode.
 data <- nhisDiseaseForcast(key, localeCode = c(11, 41), type = "Asthma", slow = T)

 # example 2 searching by localeName
 data <- nhisDiseaseForcast(key, localeName = c("\\uc218\\uc6d0"), type = "All", slow = T)

}
