% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsCorp.R
\name{npsCorp}
\alias{npsCorp}
\title{National Pension Service, Search company's information with corporate registration number.}
\usage{
npsCorp(key, regist_numb = NULL, verbose = FALSE)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{regist_numb}{character value. first 6 number of company's corporate registration number.}

\item{verbose}{logical value. If TRUE, show process bar. Default is set as FALSE.}
}
\description{
npsCorp function import data about corporation's information. See details
}
\details{
The information contains location, type, unique id and so on.
  The "id" in imported data is used at npsCorp function.
  Because the "regist_numb" argument is ambiguous(not a full corporate registration number),
  you should search target company from imported data by "name".
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- npsCorp(key, regist_numb = 208812)

}
