% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsPension.R
\name{npsPension}
\alias{npsPension}
\title{National Pension Service, Information about corporation's national pension member.}
\usage{
npsPension(key = NULL, id = NULL)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{id}{corporation's unique id which can find from npsCorp function.}
}
\description{
npsPension function import data about corporation's national pension information.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- npsPension(key, id = 22626108)

}
