% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsPensionInOut.R
\name{npsPensionInOut}
\alias{npsPensionInOut}
\title{National Pension Service, Information about corporation's national pension member in/out count.}
\usage{
npsPensionInOut(key = NULL, id = NULL, month = NULL)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{id}{corporation's unique id which can find from npsCorp function.}

\item{month}{6-digit month like 201911.}
}
\description{
npsPensionInOut function import data about the number of corporation's national pension's number.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- npsPensionInOut(key, id = '22626108', month = '201911')

}
