% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutriComponent.R
\name{nutriComponent}
\alias{nutriComponent}
\title{Ministry of Food and Drug Safety, Food nutritive components information.}
\usage{
nutriComponent(key, verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}
}
\description{
nutriComponent function import the food nutritive component information.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example 1 searching by localeCode.
 data <- drugsSideEffect(key)

}
