% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiDrugs.R
\name{quasiDrugs}
\alias{quasiDrugs}
\title{Ministry of Food and Drug Safety, Quasi-drug products information.}
\usage{
quasiDrugs(key, verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{verbose}{logical value. if TRUE, provide process bar. Default value set as false.}
}
\description{
quasiDrugs function import Quasi-drug product information like name, manufacturer, effect, and so on.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- quasiDrugs(key, verbose = T)

}
