% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoulBusCount.R
\name{seoulBusCount}
\alias{seoulBusCount}
\title{data.seoul.go.kr, count bus passenger.}
\usage{
seoulBusCount(key, day = gsub("-", "", Sys.Date() - 4), verbose = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{day}{character value. day value of YYYYMMDD type. Default is set -5 days}

\item{verbose}{logical value. if TRUE, provide process bar. Default value set as false.}
}
\description{
seoulBusCount function import the number of passenger by bus stops for a day.
}
\examples{
 key <- 'your key issued from data.seoul.go.kr'

 # example.
 data <- seoulBusCount(key, day = gsub('-', '', Sys.Date()-4), verbose = T)

}
