% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoulMaintenance.R
\name{seoulMaintenance}
\alias{seoulMaintenance}
\title{data.seoul.go.kr, realtime bike information.}
\usage{
seoulMaintenance(key, verbose = FALSE)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{verbose}{logical value. if TRUE, provide process bar. Default value set as false.}
}
\description{
seoulMaintenance function import data about apartment maintenance cost.
}
\examples{
 key <- 'your key issued from data.seoul.go.kr'

 # example.
 data <- seoulMaintenance(key)

}
