% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoulTubeCount.R
\name{seoulTubeCount}
\alias{seoulTubeCount}
\title{data.seoul.go.kr, count bus passenger.}
\usage{
seoulTubeCount(key, day = gsub("-", "", Sys.Date() - 4))
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{day}{character value. day value of YYYYMMDD type. Default is set -5 days}
}
\description{
seoulTubeCount function import the number of passenger by subway stations for a day.
}
\examples{
 key <- 'your key issued from data.seoul.go.kr'

 # example.
 data <- seoulTubeCount(key, day = gsub('-', '', Sys.Date()-4))

}
