% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_GET_content.R
\name{try_GET_content}
\alias{try_GET_content}
\title{Try httr::GET and httr::content function at least 5 times If got errors, save error message.}
\usage{
try_GET_content(url, times = 5)
}
\arguments{
\item{url}{url want to read.}

\item{times}{trying times.}
}
\description{
Try httr::GET and httr::content function at least 5 times If got errors, save error message.
}
