% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_read_json.R
\name{try_read_json}
\alias{try_read_json}
\title{Try jsonlite::read_json function at least 5 times If got errors, save error message.}
\usage{
try_read_json(url, times = 5)
}
\arguments{
\item{url}{url want to read.}

\item{times}{trying times.}
}
\description{
Try jsonlite::read_json function at least 5 times If got errors, save error message.
}
