% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_read_xml.R
\name{try_read_xml}
\alias{try_read_xml}
\title{Try xml2::read_xml function at least 5 times If got errors, save error message.}
\usage{
try_read_xml(url, times = 5)
}
\arguments{
\item{url}{url want to read.}

\item{times}{trying times.}
}
\description{
Try xml2::read_xml function at least 5 times If got errors, save error message.
}
