% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_xmlToList.R
\name{try_xmlToList}
\alias{try_xmlToList}
\title{Try XML::xmlToList function at least 5 times If got errors, save error message.}
\usage{
try_xmlToList(url, times = 5)
}
\arguments{
\item{url}{url want to read.}

\item{times}{trying times.}
}
\description{
Try XML::xmlToList function at least 5 times If got errors, save error message.
}
