% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_dataframe.R
\name{xml_to_dataframe}
\alias{xml_to_dataframe}
\title{convert xml_nodeset to data.frame, attributes as column names.}
\usage{
xml_to_dataframe(nodeset)
}
\arguments{
\item{nodeset}{result of xml2::xml_find_all function.}
}
\description{
convert xml_nodeset to data.frame, attributes as column names.
}
