% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmaLifeIndex.R
\name{kmaLifeIndex}
\alias{kmaLifeIndex}
\title{Korea Meteorological Administration, The life weather index.}
\usage{
kmaLifeIndex(key, time = seq(0, 21, 3), localeCode = NULL,
  localeName = NULL, type, slow = T, viz = F)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{time}{numeric value. The time when the data is generated.}

\item{localeCode}{numeric value. SiGunGu code which means legal area. one of localeCode or localeName should be inserted.}

\item{localeName}{character value. SiGunGu name wich means legal area. one of localeCode or localeName should be inserted. It should be Korean.}

\item{type}{character value. decide the type of index. it should be one of "fp", "st", "hi", "di", "ui", "fb", "ap", "sh" or "possible". see details.}

\item{slow}{logical value. if TRUE, give sleep inbetween importing. default is TRUE.}

\item{viz}{logical value. if TRUE, provide simple 2d visualization result. x: date, y: mean index.}
}
\value{
data.frame and visualization.
}
\description{
kmaLifeIndex function import the life weather index data last 2 days. The function also provide simple visualization using plotly.
}
\details{
kmaLifeIndex function import eight index value which calculated by Korea Meteorological Administration and related to public's life.
 Explanation about "type" as follow.\cr
 "fp" = Food poison index. It is povided from Jan to Dec.\cr
 "st" = Sensory temperature index. It is povided from Nov to Mar.\cr
 "hi" = Heat index. It is povided from Jun to Sep.\cr
 "di" = Discomport index. It is povided from Jun to Sep.\cr
 "ui" = Ultraviolet index. It is povided from Mar to Nov.\cr
 "fb" = Freezing burst index. It is povided from Dec to Feb.\cr
 "ap" = Atmospheric dispersion index. It is povided from Nov to May.\cr
 "sh" = Sensory heat index. It is povided from May to Sep.\cr
 "possible" = All index which affordable now.\cr
 See more details about each index from <http://www.weather.go.kr/weather/lifenindustry/life_jisu.jsp#>
}
\examples{
 # example 1 searching by localeCode.
 data <- kmaLifeIndex(key, time = seq(0, 21, 3), localeCode = c(4111100000, 4111156600),
                      type = "fp", slow = T)

 # example 2 searching by localeName
 data <- kmaLifeIndex(key, time = seq(0, 21, 3),
                      localeName = c("수원"), type = "possible", slow = T)

}
