% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasCandidate.R
\name{nasCandidate}
\alias{nasCandidate}
\title{National Assembly Secretariat, candidate data.}
\usage{
nasCandidate(key, election_id = NULL, election_type = NULL,
  is_pre = FALSE)
}
\arguments{
\item{key}{character value. API key issued from <www.data.go.kr>. no default.}

\item{election_id}{character value. the election id which means election date.}

\item{election_type}{numeric value. the election type.}

\item{is_pre}{boolean value. if TRUE, import pre-candidates data.}
}
\description{
nasCandidate function import profile of election candidates.
}
\examples{
 key <- 'your key issued from data.go.kr'

 # example.
 data <- nasCandidate(key, election_id = '20180613', election_type = 4)

}
